File cleanup, Proliance code working on home page
Deleted a bunch of files being used by GDOT but not by this project. Got
the Proliance code working for two values on the home page. Need to
figure out how to get the values that are in the mockup but not on the
live home page. Also need to be able to pass data from C# code to jQuery
thermometer.