Progress on #8
Created a smaller version of the region map image, updated image map.
Added a "xlg" media query break point to handle the issue of the
thermometer overlapping the region map image. Changed layout so that the
navigation goes to the bottom of the page when the page is narrow enough
to make the columns stack.