Fixed #11, #12
Moved menu item, added separate thermometers for each region to their
respective project pages. Also added all three thermometers to the
Projects landing page. For some reason, putting the thermometers in a
hidden div removes the shine effect from the bottom of the thermometer
but not for the first one that occurs in the HTML code.